+++ lib/watir-webdriver/elements/element.rb
@@ -129,7 +129,7 @@ module Watir
         end
       end
 
-      browser.checkers.run
+      browser.after_hooks.run
     end
 
     #
@@ -145,7 +145,7 @@ module Watir
       assert_has_input_devices_for :double_click
 
       element_call { driver.action.double_click(@element).perform }
-      browser.checkers.run
+      browser.after_hooks.run
     end
 
     #
@@ -161,7 +161,7 @@ module Watir
       assert_has_input_devices_for :right_click
 
       element_call { driver.action.context_click(@element).perform }
-      browser.checkers.run
+      browser.after_hooks.run
     end
 
     #