+++ lib/watir-webdriver/elements/form.rb
@@ -10,7 +10,7 @@ module Watir
     def submit
       assert_exists
       element_call { @element.submit }
-      browser.checkers.run
+      browser.after_hooks.run
     end
 
   end # Form